use Long.parseLong() instead of Integer.parseInt()

+++ xmltoexcel/src/main/java/org/assimbly/xmltoexcel/domain/ExcelWriter.java
@@ -53,7 +53,7 @@ public class ExcelWriter {
     private void setCellValue(String value, Cell cell) {
         String integerRegex = "-?\\d+";
         if (value.matches(integerRegex)) {
-            cell.setCellValue(Integer.parseInt(value));
+            cell.setCellValue(Long.parseLong(value));
         }
         else cell.setCellValue(value);
     }
